requirements.txt
playwright==1.58.0
openpyxl==3.1.5
pyyaml==6.0.2
requests==2.32.3
python-dateutil==2.9.0